[debugger] Preserve stack when adding frames.
Stack frames can get replaced when a new stop happens, and they can get
expanded on when more than the top two are needed (only two are sent
from the remote by default so exceptions can be processed quickly).

Previously everything was replaced when the frames were updated. This
would, for example, reset the inline frame hide count whtn the frames
were synced. The result is that tying "bt" in the debugger can change
the current stack frame even when the thread hasn't run.

Change-Id: I5deeb0ee6cf060856ed42f40a41e516150e1beff