[async] Change ZX_WAIT_ASYNC_REPEATING to ZX_WAIT_ASYNC_ONCE
This is a step towards removing ZX_WAIT_ASYNC_REPEATING.

As part of this change, remove port_wait_repeating(), which is a direct
consumer of ZX_WAIT_ASYNC_REPEATING, and instead rely on port_wait().

Change utests not use ZX_WAIT_ASYNC_REPEATING.

This is a reland. The problem with the first version was that it
could often call zx_object_wait_async() more than once on the same
timer and port in the async_loop code, which results in multiple
callbacks with no work, and those would continue to register
even more object_wait_async's.

ZX-3090

Test: CQ
Test: modular_benchmark_story, the failing peridot test now passes.
Change-Id: I820babeb1bc0dbce5de471b84d7ea76cb53a908c