[upci] Change bus scanning to DFS
Per the PCI Local Bus Spec, bus scanning and id assignments
should be done in a depth-first manner. BFS was easy to implement
for bring-up, but this change makes the driver spec compliant. Rather
than have recursion in a driver, the scanning is implemented by using
BusScanEntry structs to keep track of where we are in the scan via
a traversal stack.

ZX-3146

Change-Id: Ib919a97a2f008a51ebafc5a2362f1e153c5e4cd7