[debugger] Scope types with abstract origins.
Previously the enclosing scope of inlined functions would be incorrect
when the function lacked a separate declaration. Inlined member
functions (what our tests covered) always include a declaration, but
inlined non-member functions don't.

When the declaration was omitted, the result was that the name of the
inlined function looked like it was nested inside of the function it was
inlined into.

This uses the correct namespace for this case. The enclosing namespace
should come from the parent ot fhte abstract origin if there is one. The
decoding of the parent of the abstract origin is hooked up to all types.
So far I have only seen abstract origins for inlined functions, but
theoretically they could exist for other types so this is mostly for
completeness.

Change-Id: I7ec463a487c6f5074be3dcbb9cecf4f6b1b4c777