[zx][devmgr] Create service for devmgr debug dumping.
Device manager will export a service fuchsia.device.manager.DebugDumper
and svchost will list it in public /svc/ and proxy incoming requests
back to devmgr.

This service will handle the dumping of internal device coordinator
state to text for debugging purposes.

The old approach, that went through dmctl, used a socket to get the text back to 'dash'. I didn't
like that because it was implicitly depending on the socket buffer being
unbounded as dash would perform a synchronous fidl call then when that
returned it would go read the socket. If/When socket buffers become
bounded this could cause a deadlock in dash and devmgrs asyncloop
thread.

I also considered returning a string in the fidl interface but
since this is a simple interface it means that string would need a set
size. In this approach it is up to the caller to determine how large of
a buffer to provide and doesn't require fidl changes if they want
more/less, we could also provide a flag to allow callers to dynamically
get a bigger buffer.

Issue: ZX-3403 # Break out debug dumping service.
Test: ran {dump, drivers, devprops} on "dm" in qemu.
Change-Id: I71fbb22b587c235b329c02c64a829c034c27e34a