[amber] Rewrite OTA test to use golang ssh client
This refactors the OTA tests in order to use the golang native
ssh client, instead of expecting that the ssh executable is
available in the environment. In addition, this does a minor
amount of reorganization to extract out a client abstraction
for the tests to interact with the remote device.

Test: Verified OTA test still works.
Bug: PKG-500 #comment migrated to go ssh client
Change-Id: I841955dd396052557a0ef32e7415ba31997e0e88