[inspect-vmo] Fix use after free bug in Object.
FreeObject was setting the internal state_ member to null when freeing
an object, but if that Object contained the last reference to the state,
the state would get deleted before it was done being used in the context
of FreeObject.

This change makes Object itself responsible for cleaning up its own
state.

TEST=runtests -t inspect-vmo-test

Change-Id: I51ff541f2834deeab24dfaeaff8c0617b77f3775