[elflib] Add method for getting PLT offsets from ELF files
We can get symbols out of the static or dynamic symbol tables, but for
PLT entries they tell us nothing about the actual location of the
object, even after loading. We're also particularly interested in PLT
locations, since there are times when we can set breakpoints in the PLT
but not the target method.

The approach to getting them is architecture-specific and involves
unpacking information from instruction arguments in machine code. Right
now we only implement for x64.

Change-Id: I7530347f669838b4ddf7b90c3f041ad7b8798187