[build][set] Remove petal/layer guessing logic from 'fx set'
This removes the logic used to guess the current petal/layer for
determining default board/products.  Instead, to preserve current
developer expectations, this sniffs the top-level jiri manifest to
determine if a vendor repository is in use and checks in that vendor
repository for configurations.  If there is a vendor repository
specified and the requested configuration is not there, or if there is
no vendor repository, this looks only in the top-level configuration
location (either //boards or //products).

Configurations in non-canonical locations can still be specified but
only by full path, not short name.

Change-Id: I842e3bb43649afaec73aabf6df6968daea8aff52