[debugger] Enhance get command.
Support nouns for the "get" command, so one can type "process 2 get" or
just "process get" to get just the settings for the process.

Allows searching for a specific setting on a specific noun via "job get
filters".

Outputs the noun index for the setting summary, so "get" will print
"Process 1" and "Thread 5" (for example) so it's clear what the returned
setting context is.

Generalizes the retrieval of settings to avoid code that manually
checks different object types. Most of this is stored in an arrays with
one helper to get the right setting schema and object for a given type.
This should make the settings more extensible in the future.

Adds a "global" noun to the command syntax. This allows specification of
the global settings store via "global get". This noun will also be
needed in other contexts.

Removes printing of the settings schema when no object exists.
Previously "get" would fall back to printing the schema. when e.g. no
thread exists. But this gets either weird or complicated when you have
to support "thread get foo" on the schema. Since we don't have any
settings in this category, it's not possible to test and it's difficult
to tell how important this will be. We can add it back later if we run
into a need for it.

Change-Id: I3a40975946b8fc483edaca7335fbe80d13877c75