[benchmarks] Re-enable test executables that use the perftest library
These got broken/disabled by Zircon's switch to using GN to build.

Fix Zircon's perf-test executable to be given its original name,
"perf-test".  When output_name wasn't set, GN was assigning it the
default name "perftest-test".  That meant that the logic in
garnet/tests/benchmarks/BUILD.gn for copying it into the
garnet_benchmarks package was failing to find this executable and
silently failing to copy it.

Bug: INTK-858
Test: garnet-x64-perf-dawson-canyon trybot
Change-Id: Ia3455f7506bec873774f844e04b99ea1611721c7