[debugger] Fix inline stepping bug.
Removes an inline step condition that didn't make any sense.

The exact bug is if you had this code:

  InlineOne(); InlineTwo();

Stepped into InlineOne(), and then used the step over controller to step
out of InlineOne(), the source code will show being in the first
instruction of InlineTwo() rather than about to call it. This is because
the deleted call to the step controller would step the program.

Deletes the test that was covering this behavior. The problem was this
test was not setting up enough symbol information for the code to make
the correct determination that the line was the same after finishing the
function call. The result was that the later calls to the step
controller would not kick in when needed, requiring the now-removed step
call above to succeed.

I considered fixing the test but the fixed test would look almost
exactly like the one below it except with a physical frame instead of an
inline frame. Assuming the Finish controller can finish both physical
and inline frames (it has its own tests), the code executed by the step
over controller will be identical.

Change-Id: I32eb699c7ed204bb6d33ccf6c8d1f9e192eca5d0