[runtests] Add dry run mode to device runtests.
Add a -d flag that will just print what tests would be run to the
device's runtests command.

This is valuable as an easy way to tell what tests the device has
without running all of the tests and trying to look through the result
to see if the test you care about is in there.

Test: ran it.
Change-Id: I427460a082570e54df94538c40b5564a270efa33