[rust][fuchsia-trace] Duration events auto-drop.
Consolodate the |DurationScope| and |DurationDropScope| into a single
type that always sends the DURATION_END event when the |DurationScope|
is Dropped. Additionally, stop deriving Copy/Clone to ensure that only
a single DURATION_END event is sent for each DURATION_START.

If precise control over the duration begin and end is required, you can
use the `duration_begin` and `duration_end` macros.

Ex, the following are equivalent (however the former makes more
efficient use of the trace buffer):

  {
      duration!("foo", "bar", "arg1" => 2, "arg2" => 3);
      ...
      // event is recorded here.
  }

  {
      duration_begin!("foo", "bar", "arg1" => 2);

      duration_end!("foo", "bar", "arg2" => 3);
  }

Test: fx traceutil record
Change-Id: I2b105d6ec4817b5a9c54d71768f8ea488fe58535