[elflib] Make MemoryAccessor internal
The MemoryAccessor interface was designed to give users complete control
over how memory was managed when we fetched data from ELF files and
address spaces. However, the use cases are becoming better-defined, and
we can now expose a simpler interface that meets our uses specifically.

We remove the public memory accessor in favor of an overloaded create
function that allows several ways to specify how the data is accessed;
either a callback, a bare memory region, or a file. We cache reads in
some of these cases and do not cache in others. We can also specify
whether to use file offsets or runtime addresses when looking for data
structures. The result is we retain most of the flexibility of the
previous interface but shed most of the complexity.

Change-Id: I6b63e1c8c97082b93be6433d5e3c8a1c77506573