[zircon] Fix the missing dependency .fidl dependency in depfile
gen-fidlc-rsp.sh reads the response file and looks for input .fidl files
that follow the --files argument. The problem is that handle_args keeps
local state, concretely whether it saw --files argument, but
read_rspfile is iterating over the response file lines and invokes
handle_args for each line, so the local state is lost, the .fidl input
file is never detected and hence never mades it into the depfile which
means that the action is never re-run on .fidl file changes as a
consequence.

Bug: BLD-367 #done
Change-Id: I7c8b0e4abef8dafeaa7cd4ed33a283b93d09bbf4