[guest][vmm] Drain closed vsock connections
Previously the socket/channel connection code would send a SHUTDOWN command to
the guest when it saw PEER_CLOSED, even if there was still data in the connection.
This change prefers to always try and send any remaining data the client may have
put in the connection, before sending the SHUTDOWN.

FLK-65 #done

Test: Ran the VsockGuestTest.ConnectDisconnect test a few hundred times in a release
build. Without this change it would typically fail within 20 runs.

Change-Id: I11081b4ba7758aceed988d3f70d9de1805de3d1b