[debugger] Fix crash on job attach failure.
If attaching to a root job failed, the agent would crash.

The problem was that a zx_status_t was being passed through a bool which
would corrupt the value (force it to 0/1). This was fixed.

The value of 1 would cause the zx_status_t string converter to return
null which would confuse the printing code. This has been changed to
return "<unknown>" for future cases.

Change-Id: I2383a25590471ffec0db09de4d3484146582b61e