[bt][bt-gap] Output top-level bt-gap errors to syslog
When bt-gap errors and returns Error from the top level function, log
that error to syslog before returning it from the process. This aids in
diagnostics if another component requires a service from bt-gap, and
bt-gap fails during execution, as the error will now be easily visible
by reading syslog.

test: Run bt-gap in a situation where it will fail (For a motivating
example, in a hermetic testing environment where the Stash service is
not avaialble), and then run `$ fx syslog` and check for the error
output.

Change-Id: Ic1f0320289f15f36b96f4b2f0db9f34e90c7820b