[debugger] Resolve paths in ids.txt relative to containing dir
The ids.txt files we see today tend to have absolute paths only, but
relative paths are possible. These paths can either be relative to where
the generating tool was run, or relative to where the ids.txt file was
output. The first is hard to do anything useful with so we just always
assume the second.

Bug: DX-1074
Change-Id: Ie7dc305eede1b552130dc821d0815da9bfee8f1e