[scenic] Make real vsync signal reach frame scheduler
Vsync timings and intervals are currently indirectly estimated by the frame scheduler through the display, and only updated when there is a frame to present.
This CL changes it so that the display gets the real vsync signal and is responsible for keeping track of timings and intervals.

SCN-563 #done

Change-Id: Ica705b79b2aa3def87f66db1b19bbd167b0778a9