[mediaplayer] rationalize distinguished PTS values
mediaplayer uses three distinguished PTS values: a null/unknown
value and the minimum and maximum non-null values. The null value
is defined in the fuchsia.media FIDL library as NO_TIMESTAMP as
the max<int64> value.

Previously, mediaplayer defined Packet::kUnknownPts, which is
supposed to play the same role as NO_TIMESTAMP. It was defined
to be the min int64 value.

This change replaces Packet::kUnknownPts with Packet::kNoPts, which
has the value max<int64>. A static assert checks that this is the
same as NO_TIMESTAMP. Most uses of NO_TIMESTAMP in the code have
been replaced by kNoPts.

Packet also defines kMaxPts, which is max<int64>-1, and kMinPts,
which is min<int64>, which are used in a couple of places.

TEST: no behavior change...tested media playback for regressions

Change-Id: I994c4e0d49d5dbfeeaf04ca0ef38f8fd92528d0d