[wlan][ralink] Tx Report Routine Should Not Kill Interrupt Worker
The original code checks the packet tx info provided from ralink
hardware. If the info is spurious, the code would warn that.

Instead of returning error, ignoring that stale info is a better
idea. Otherwise, it will kill the interrupt worker, which is also
used to update the hardware beacon info.

BUG=WLAN-992
TEST=With my E2E test, on Toulouse ralink and Ava ath10k.
     ~/wlan-e2e-test/reliability_test_ap_role.sh

Change-Id: I07a313799c59930e5591736bfdeab1f4e9bdc66b