[debugger] Arbitrary notifications out of the client.
There are a lot of events that do not occur as part of the normal
"notification" flow of the debug agent (eg. disconnected).

We need a generic way of telling the UI that something happened. Since
the session is the most overarching object the UI connects to, it seems
reasonable to have an observer on any events it detects.

This has the advantage of having the output be in sync with the terminal
and avoid the whole "staircase" nonsense when outputting directly to the
terminal.

Added a disconnected message and ported the previous error messages.

TEST=manual

DX-952 # Comment

Change-Id: I85da021010f6cca225ac25e711d5364d1cb522f7