[linenoise] fix getCursorPosition and update
This patch merges in changes from upstream, which contains a few small fixes.

An additional patch is introduced that replaces the error result from
completeLine with a non-result, a repeated read will either detect the error
state or carry on. This avoids int to char conversion.

Most Fuchsia patches are ported forward, except for the patch that uses
`poll`, as that has been prematurely timing out reads in getCursorPosition -
which is unsafe, as the terminal later reads the cursor position response as
if it is user input, sometimes resulting in fatal thrashing of the prompt in
a loop (which can be caused by a racing read particularly in concert with
being run in terminals that are attempting to e.g. report focus events).

Change-Id: I83a9eb820c0ac07a3fdbee2526c3a7cbacae8eff