[bootsvc] Use FIDL to transport root resource
Use a FIDL service to transport the root resource between bootsvc and
the next process in the boot sequence, which is currently devmgr.

The FIDL service is currently provided under /bootsvc in order to not
collide with /svc. This will change in the future once componentmgr is
introduced into the boot sequence.

ZX-3385

Test: Added to bootsvc integration test, and ran DDK tests.
Change-Id: Id6f8219c971a4dc9958040a1a56a05d718e83eac