[debugger] Fetch registers from inline frames.
The FrameSymbolDataProvider checks to see if a frame is topmost to
determine whether the thread's current registers are valid. This broke
when we added inline frames which share the same registers, but may not
necessarily be topmost.

This change will make the current CPU registers available to the top
physical frame and any of its inlines.

TEST=step into an inline function. "p rax" (for x86-64) should work for
both the inline frame and the top physical frame.

Change-Id: Iae5870451e452fc0592602fa4a71fb6d9281fae8