[netstack] ensure we don't miss events
Register to receive event notifications before starting to service
socket control operations. This prevents the race where the endpoint is
ready before we register for events, and we deadlock because we never
receive the notification. In one scenario, it was possible for a socket
to start listening and receive an incoming connection before the
loopRead goroutine was scheduled; that would result in loopRead
deadlocking as it waited for an event notification that would never come
(because the endpoint was ready before registration). All this was
observed in...

FLK-71

Change-Id: Ia985f71d23729608a35a240c5672bb9dd658fdb0