[debugger] Parse function calls in expressions.
This parses function calls in expressions like "print Foo()" or
"print reinterpret_cast<Foo*>(bar)" (where the function is
reinterpret_cast in this case).

Currently these are not hooked up to anything and will issue an error if
one is ever executed. This will be used to implement reinterpret_cast
and static_cast in the future.

Fix a problem in the template parameter list parser that allows trailing
commas. Test added for this case.

Change-Id: I24299fbd3e8242b778c6dfffe8cbcc1d9e66b9f9