[pseudo-fs] Require DirectoryEntry to be Send.
Apparently if we do not do this explicitly, `Send` is not automatically
inferred and that makes it impossible to construct a pseudo directory
tree and then send it to another thread for execution.

Seems like this has a drawback in now requiring all the callbacks to be
implemented using only `Send` types, rulling out, for example,
`RefCell`s.  I am not sure if this would be a considerable inconvenience
for usage in actual applications. It would be great if we could make
this generic over `Send`, but I do not understand yet how to do it.

Change-Id: I4543c33f4ba12011f9a91847736fdb21205f134a