[amber] Rewrite the System OTA tests to download build artifacts
This rewrites the tests to download the package repository
from a given build id using the tools lkgb and artifacts.
To run the tests, simply run:

```
% ~/fuchsia/out/x64/tools/system_ota_test \
  -zircon-tools-dir ~/fuchsia/out/build-zircon/tools \
  -builder-name fuchsia/ci/fuchsia-x64-release \
  upgrade
```

It will attempt to OTA to the latest fuchsia-x64-release,
and check if the update is successful by checking if
/config/build-info/snapshot matches the version we downloaded
from the build artifact repository.

Test: Verified OTA tests can verify the OTA succeeded.
Bug: PKG-511 #done
Bug: PKG-512 #done
Change-Id: Id35da356003e9690fa0c98fe38226b253e7f97b6