[botanist] Move syslog streaming to botanist
This moves syslog streaming from the testrunner to botanist.

Why:
(1) now we can retrieve syslogs even if the testrunner subprocess - or
the test task itself - times out. (Needed in the case of long-running tests
that hit swarming's streaming maximum; the syslog will be a complete
record of what happened and be invariably present.)
(2) this lessens reliance on the runtests model of including syslog.txt in
the `ouptuts` of summary.json.

Change-Id: I35951e31be94d257a70cff5449b9096e7d80ed45