[build][shell] simplify shell command generation
We know almost all of the information at gen time, so produce the gen
artifacts and manifests all at gen time. This fixes a null-build issue where,
because shell-commands depends upon all other build targets, and the
"create-shell-commands" target was not producing a depsfile, it was becoming
dirty after a complete incremental rebuild of any other package.

Change-Id: Ia85fd6110a32ef3e5105a00ddfd48d66d674bfd9