[debugger] Hook up reinterpret_cast.
Adds support for reinterpret_cast for a few hardcoded types. This
implementation of reinterpret_cast is very permissive and supports
converting some types that C++ will not convert, but which is useful to
have in the context of user typing in a debugger.

Real type parsing for the current process' symbols will come in later
passes.

Change-Id: I86227318c02e1cf32746b8c15b78d7726f1ba04f