[sdk/vfs] Close connections if describe fails.
The description in io.fidl says the connection closes if Describe fails,
but this does not happen in practice. If the output of describe is not
set we just see a FIDL encoding error in the logs and the connection
hangs.

This change actually closes the connection if we are about to attempt
responding with an invalid NodeInfo union.

TEST=follow up CL triggers this

Change-Id: I3064b4e368de6138b2a27c8d903af911e6459af3