[wlan][ath10k] Enable HT Rates for AP Role.
Apparently the ath10k requires QoS is enabled to send out packets
at 11n+ rates. In this CL, the MLME will mark the client supports
QoS and let the driver/firmware to decide what to do.

BUGS=WLAN-767,WLAN-108
TESTED=With my Toulouse (as AP) and Ava/Pixel 2 (as client).
       The AP can send out packets at 11n rates.
       Also regressed the ralink and broadcom drivers to ensure
       they are not broken.

Change-Id: Id75ce89c7717e759d8e4d405e94ae27d1ac38335