[mediaplayer] support no-pts audio packets and discontinuity
This CL adds support to mediaplayer for packets without PTSs in audio
pipelines. The discontinuity bit is also plumbed through. This is
required to handle bluetooth frame drops properly.

The Packet class has been modified to carry the discontinuity flag,
which is used in unstamped packet stream to indicate that the marked
packet need not be presented immediately after the previous packet.

The SimpleStreamSink implementation and the audio decoder have been
modified to plumb through the discontinuity bit.

The audio renderer node has also been modified to assign PTSs to
unstamped packets using its understanding of the behavior of the
actual FIDL AudioRenderer and discontinuity bits. In theory, the
FIDL AudioRenderer should handle this itself, but MTWN-243.

TEST: file playback for regressions, bluetooth sink

BT-723 #comment

Change-Id: Ic204a367f842969e2508fb95f1ce60f40a8d75cb