[kernel][sched] Fix idle/run time accounting in fair scheduler.
CPU stats and thread_runtime() expect a specific behavior for
thread_t::last_started_running. When the thread is runnable the
delta between this member and current time is added to idle time
or runtime, respectively. Fix the fair scheduler to respect this
relationship.

Also perform some minor cleanup of dead code and avoid computing a
superfluous new time slice when switching to the idle thread.

Bug: ZX-3615
Test: kstats -c produces load times <= 100.00%.
Change-Id: Ibc5e477426fac75965bb2dbdb412ce16ffa29e2c