[artifacts] Refactor to enforce a new GCS layout
We need to use this package to upload test logs from the
testrunner, and to locate them when uploading data to
ResultStore.  This change updates the artifacts package
to adhere to a specific layout, so that all CloudStorage
clients can consume this package (through some CLI or as
source) and write or read to and from Cloud Storage in
the exact same way.  We will update our infra recipes to
use this library to upload to GCS.

IN-888 #comment

TEST=ran the commands in artifacts/main.go by hand

Change-Id: If9b9e5443d5eebb733c295104cdab52d3d34ad09