[garnet] Fix WeakPtr move constructor, and cleanup garnet headers
WeakPtr.h is the only "code" change. The code implements a custom move
constructor, but I believe the implementation is not best practice. It
was leaving the _ptr of the moved from object set to the original value.
If someone happens to use the moved from object (even though they
shouldn't), it might work, even though it's no longer participating in
the reference counting.

Other files are only removing make_copyable.h includes.

Test: No behavior change
Change-Id: I065126b88db4c791f1fb190ccf50b8055d9c1bab