[devmgr] Allow closing device instances opened via device_add
When calling |open| or |open_at| on a device it may choose to return a
new device instance and open a connection to that device. When the
connection is closed the device is removed. The mechanics of this are
that the devhost claims the reference to the device that is leaked by
|device_add|.

A connection to a device instance may also be created by |device_add|
when passing a |client_remote| to the devmgr. This behaves almost
exactly the same as a device created in |open| or |open_at| except that
the device persists after its closed. This CL addresses that
inconsistency by not leaking a reference to the device in |device_add|
for this specific case.

Change-Id: I2c2fc82a9059e32c46f72d40b3131babc2d5f9cc