[persistent-storage] Fix partially broken whitelist mechanism
Added the dependency in the wrong place in fxr/254330 when first
implmenting the whitelist for 'deprecated-global-persistent-storage',
which meant that a package target only had to set the right variable and
the visibility whitelist was never properly checked.

Also cleans up two whitelist entries that are no longer needed (targets
migrated in fxr/261678 and fxr/261550).

Change-Id: Ie8b5cefba0eb26cf6bb8c0b03095f73206916c85