[peridot] Initialize modular Operations with make_unique
OperationContainer classes are designed to "own" their Operations, and
the private "Hold()" method hides the fact that each given Operation* is
converted into a unique_ptr, and later destructed by the container.

This change makes the ownership more explicit by requiring Operations to
be "moved" into the container, via unique_ptr.

In addition, it was noted that the practice of including the macro
FXL_DISALLOW_COPY_AND_ASSIGN is discouraged, and no longer permitted in
Google code, and that the Operation base class already deletes the copy
constructor and assign operator, so there's not much point in
re-declaring them in derived Operation subclasses. So the macro is
replaced by the explicit (and now public, as per Google style guide)
deletion of these methods, in operation.h, and elsewhere, Operation
subclasses no longer include the macro at all.

Test: No behavior change
Change-Id: Ib699d643e3b96d7f0c2857abcb32668f7abe747d