[fidl][go] Go bindings soft migration: add DispatchNew method to stubs
The Go bindings export the `Stub` interface, which bindings
implementations must implement. Currently, the `Dispatch` method uses
the soon to be deleted `Payload` type. In order to have a soft
migration, we need to:
- Implement both `Dispatch` and `DispatchNew`
- Then we can land the change in third_party/go
- Post-migration, there will be cleanup steps to remove the suffix
`New` from newly introduced method.

See next CL on third_party/go
https://fuchsia-review.googlesource.com/c/third_party/go/+/262128
which introduces this method on the Stub interface

Change-Id: I64ebe276c5626c1406d4628f18686dad41220679