[debugger] Write a new number parser.
This will replace the number parser in console/command_utils.cc, the
number code in the parser, and the number code in the ExprNode in a
future patch. Currently it is not hooked up to anything but tests.

This handles most input cases for C++ including octal, decimal, and
binary prefixes. Most of the complexity comes from generating ExprValues
of the right types. It uses a simplified version of C++'s rules to
generate a value of the correct type for a given input.

Change-Id: I78ac6a400d7fab7c2dd71a81595e436b8127d2b7