[vim] Use the root file type for FIDL goldens
There are a bunch of FIDL tests that compare generated files against
"golden" versions. The golden versions are previously generated checked
in versions so that when changes to code generation occur we can
validate that they are what we intend.

The golden files have ".golden" appended to their name which means VIM
can't work out how to highlight them. This causes VIM to execute the
hooks it would invoke without the ".golden" extension when any *.golden
file is loaded. Most importantly syntax highlighting works.

Change-Id: I42fe556599ff594d4d0d219bd6822459c285f79f