[devmgr] bypass dmctl for dm {reboot,shutdown,suspend}.
Export fidl service to handle reboot, shutdown, suspend "dm" commands.
This pushes the distinction between these (what flag is used) to the
dash command handler and keeps the interface as one call.

Eventually this can evolve to be multiplexed to other things that will
care about this (some power services perhaps) so devmgr doesn't need to be
responsible for shutdown, it can just be informed about it.

Issue: ZX-3403 # handle reboot,shutdown,suspend.
Test: ran commands in qemu, appeared to work as expected.
Change-Id: Ib63718a35c82a154bdce5bb834867e46956223a1