[debugger] Use enum for scope visiting result.
The class hierarchy visitor previously used a boolean result code which
could not disambiguate success and failure, and made it difficult to map
the boolean result to what happened.

Using an enumeration for this case makes it easier to read and can
disambiguate failure cases. This will be important as more visitor code
is added.

Change-Id: Iac190ebd46827bdc9f66af36386ff72b4e954df6