[debugger] Convert dwarf tags to an enum.
Previously the symbols used constants for DWARF symbol tags. I was
originally thinking this would be better since the symbols can generate
other values that aren't strictly in the enum.

But so far this has not helped, it makes it less clear what a "tag" is,
and the Symbol class declaration is difficult to read.

This moves the tag definitions into an enum in a separate file. The
naming of users is updated.

There should be no visible change.

Change-Id: Id99ba97fa9dc09e90d6ef3df902f2de04b8912ae