[debugger] Move type querying to dwarf_tag file.
Moves the modified type querying function to dwarf_tag.cc, and adds a
new function to test whether a tag is a type. This is used in an
assertion in the Type object.

Adds a little bit of missing tag documentation.

With the exception of the added assertion, there should be no
user-visible change.

Change-Id: I441ab4d3c871efdd0f2047f5b6caf4edc9255374