[media/mediaplayer] move Metadata and TimelineFunction, phase 1
This CL is the first of 5 moving fuchsia.mediaplayer.Metadata and
fuchsia.mediaplayer.TimelineFunction to fuchsia.media. Both of these
structs are media-generic, and fuchsia.mediasession needs to
reference them. These definitions are referenced in topaz, hence
the multistep transition.

In this CL, we add the FIDL definitions in fuchsia.media, deprecate
the definitions in fuchsia.mediaplayer, and add the fuchsia.media
versions as alternates in a couple of places.

fuchsia.mediaplayer.SourceManager.CreateStreamSource has a parameter
of type Metadata?. Its only usage passes a null, so we don't need to
soft-transition that method.

TEST: no behavior change
Change-Id: I7c037d7523ee453b19103439c8d93f6cce8df426