[async] Incorrect condition in ASSERT triggered in async code
A race condition exists when a timer fires and enqueues a packet on a port,
and the run thread picks it up with zx_port_wait and then does not reach
the dispatch code. Meanwhile another thread comes in to cancel the task
that just fired. In this case zx_port_cancel() returns an error which should
not be regarded as a real failure.

Bug: WLAN-1001 #done

Test: CQ.
Change-Id: Ic687c9a5a1080fe546cae5641788cf5714cfc6f4