[debugger] Handle nullptr_t and typedefs
This is encoded in the symbols as an "unspecified type" and requires some
special-casing.

Follows typedefs when printing. Previously it would print an error in
these cases.

Decode unspecified types and imported declarations. The imported
declarations ("using std::vector;") needs some additional infrastructure
to even encounter, so these are not fully complete or tested yet.

Change-Id: Ie141dab78981a81aaadf496ea8b40ecf0d76a91a