[media/mediaplayer] move Metadata and TimelineFunction, phase 3
This CL is the third of 5 moving fuchsia.mediaplayer.Metadata and
fuchsia.mediaplayer.TimelineFunction to fuchsia.media. Both of these
structs are media-generic, and fuchsia.mediasession needs to
reference them. These definitions are referenced in topaz, hence
the multistep transition.

In this CL, we change the original references to fuchsia.mediaplayer
Metadata and Timeline to instead reference fuchsia.media. This can
only happen when topaz has stopped using the fuchsia.mediaplayer
fields (phase 2).

The orignal fuchsia.mediaplayer definitions are deleted in this CL.

TEST: no behavior change

Change-Id: Iaac725bfcfc4a665e3884afbfc4163fbcb4e3771