[media/mediaplayer] move Metadata and TimelineFunction, phase 5
This CL is the fifth of 5 moving fuchsia.mediaplayer.Metadata and
fuchsia.mediaplayer.TimelineFunction to fuchsia.media. Both of these
structs are media-generic, and fuchsia.mediasession needs to
reference them. These definitions are referenced in topaz, hence
the multistep transition.

In this CL, we remove the alternate fields added in phase 1. In
order for this to work, topaz must have transition to the original
fields (phase 4), which were previously changed to the new
fuchsia.media types (in phase 3).

TEST: no behavior change
Change-Id: I2303769962298373a5fce23d4ad23b3a016ad872