[system-ota-tests] Parameterize ota tests to target multiple devices
This parameterizes building and running the system OTA tests
in order to ease configuring it for different devices. With
it, new devices can be configured to run all the OTA tests
by simply calling:

```
system_ota_tests("...") {
  release_builder_name = "fuchsia/ci/fuchsia-x64-release"
  environments = [ nuc_env ]
}
```

`system_ota_tests` works by generating a shell script that
just calls out to the actual test file, but passes along
the builder name. Then we register this shell script as a
test spec with a unique environment label and service account
in order for it to run on isolated hardware, and have the
ability to download build artifacts.

In addition, I also needed to do a few things outside of the
tests:

* I extended go_build to allow me trick go_binary to
  build the test as a test file, but not to register it as a build.
* I taught the test_spec action to be aware of the system_ota_tests
  label.

Bug: PKG-575 #done
Test: Confirmed tests pass on device
Change-Id: Ic6e60c2c729cc44478c4d430fb6f2a4ffddad5b5