[input_reader] Iterate over all HID Reports
Iterate over all HID Reports in the Report
Descriptor and create a Device for each
Report ID that requires its own device.

This adds support for composite devices
that have multiple Reports. E.g some
touchscreens have seperate reports
for the touchscreen and the stylus.

Only fail if we are unable to find
a single HID Report that we can
parse correctly.

SCN-1251 #comment

Change-Id: Ia366232fc243ec634365615e16e3f48c68c12d94