[debugger] Separate out name parsing.
Separates out name parsing into its own function in the expression
parser. Previously this was done as part of normal parsing, and name
handling was split between the name, "::", and "<" token handlers.

The split out logic makes it difficult to make determinations about
what to do with a name. For example, if it's a type name, the parser
needs to look for things like "*" after it that are part of the type
name (this is still unimplemented). This is very difficult to do when
name parsing is spread around.

This also adds support for parsing expressions differently depending on
whether things are types or not. For example the parsing of
"Foo<bar>baz" depends on whether "Foo" is a template or not. The symbol
system is not hooked up yet, but there is now a callback that can be
implemented in a future patch.

Change-Id: I9043f9e29fe72913168a5fb40932d482ec009ccf