[media][e2e] Move virtual_audio driver from tests to prod
Although the virtual audio driver is a software-only construct, for
now all drivers are associated with a board, as opposed to a product.
Accordingly, this CL moves the virtual audio driver from
garnet/tests/virtual_audio into garnet/prod/drivers, which is imported
into boards/x64.gni and boards/arm64.gni and by so doing, includes all
available drivers in these generic baseline hardware profiles.

If/when we adopt a mechanism for the virtual audio driver package to
be demand-installed, and for the driver to be demand-loaded, we should
refactor this arrangement.

Test: build, CQ

Change-Id: I818c2aa71a9fb933ceefaa5372665c09752688c5