[zircon][hid] Add Insert functions to Report
Add the ability to insert values into a Report.
Either raw bits can be inserted into the Report,
or the value can be converted into the correct
units.

These are the opposite of the Extract methods.
Calling the Insert and it's corresponding
Extract method should give back the same value.

Change-Id: I5f1b7f5bb89ebecbfb4ffb548577ffd11917dab7